+++ src/config.js
@@ -3,10 +3,10 @@ const config = {};
 config.version = {
     app: {
         major: 1,
-        minor: 0,
+        minor: 1,
         patch: 0,
     },
-    build: 54,
+    build: 6,
 };
 
 config.googlePlacesAPIKey = "AIzaSyAbqz9XQOVIFRtl6X1sovgna3SHDJHtKM0";
@@ -14,5 +14,7 @@ config.googlePlacesAPIKey = "AIzaSyAbqz9XQOVIFRtl6X1sovgna3SHDJHtKM0";
 config.corks = {
     visited: 5,
 };
+
+config.development = true;
 
 export default config;