+++ CHANGELOG.md
@@ -7,7 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
 
+### Changed
+- Development node to database
+
+### Removed
 
 ## [1.0.0] - 2018-02-15